+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 27.0.0
+## 27.0.0 (October 9, 2021)
 
 ### Enhancements
 - **Forum Changes**